pyproject.toml
[project]
name = "apex-risk"
version = "0.1.0"
description = "Live Risk Management System for options and derivatives trading"
authors = [
    {name = "Chenxi Li", email = "chenxi.li08@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}

dependencies = [
    # Core dependencies
    "ib-async>=2.0.1",         # Interactive Brokers API client (maintained fork of ib_insync)
    "pyyaml>=6.0.1",           # YAML configuration parsing
    "pytz>=2024.1",            # Timezone handling for market hours
    "yfinance>=0.2.36",        # Yahoo Finance for fallback market data and beta values

    # TUI and database
    "textual>=6.11.0",         # TUI framework for terminal dashboard
    "asyncpg>=0.31.0",         # PostgreSQL async driver for persistence

    # Data science stack
    "numpy>=2.3.0,<2.4",       # Numerical computing library (capped for numba/vectorbt compat)
    "pandas>=2.3.3",           # Data manipulation and analysis
    "pyarrow>=15.0.0",         # Parquet file support for historical data storage
    "scikit-learn>=1.8.0",     # Machine learning library
    "scipy>=1.11.0",           # Scientific computing (for DSR calculations)

    # Trading calendars and backtesting validation
    "pandas-market-calendars>=4.3.0",  # Trading calendar support (NYSE, etc.)
    #"pypbo @ git+https://github.com/esvhd/pypbo.git",  # Probability of Backtest Overfit (CSCV method)
    "duckdb>=0.10.0",          # Embedded analytics database for backtest storage
    "pydantic>=2.5.0",         # Data validation for backtest specs

    # Trading and backtesting
    "backtrader>=1.9.78.123",  # Backtesting framework
    "futu-api>=9.4.5408",      # Futu OpenD broker integration
    "TA-Lib>=0.6.8",           # Technical analysis indicators
    "vectorbt>=0.28.2",        # Vectorized backtesting and analysis

    # Optimization
    "optuna>=4.6.0",           # Hyperparameter optimization framework

    # Verification
    "jsonschema>=4.20.0",      # JSON Schema validation for contract verification
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",           # Testing framework
    "pytest-cov>=4.1.0",       # Coverage reporting
    "pytest-asyncio>=0.21.0",  # Async test support
    "hypothesis>=6.100.0",     # Property-based testing
    "mypy>=1.7.0",             # Static type checking
    "black>=23.12.0",          # Code formatting
    "isort>=5.13.0",           # Import sorting
    "flake8>=6.1.0",           # Linting
    "bandit>=1.7.0",           # Security linting
    "safety>=2.3.0",           # Dependency vulnerability checking
    "pre-commit>=3.6.0",       # Git pre-commit hooks
]

observability = [
    "opentelemetry-api>=1.20.0",                  # OpenTelemetry API
    "opentelemetry-sdk>=1.20.0",                  # OpenTelemetry SDK
    "opentelemetry-exporter-prometheus>=0.41b0", # Prometheus exporter
    "prometheus-client>=0.17.0",                  # Prometheus client library
]

# Full installation includes observability
full = [
    "apex-risk[dev,observability]",
]

[project.scripts]
apex-risk = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]
# Explicitly include directories that might be filtered by .gitignore patterns
[tool.hatch.build.targets.wheel.force-include]
"src/domain/signals/data" = "src/domain/signals/data"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Default excludes
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | legacy
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["legacy"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["legacy/"]

[[tool.mypy.overrides]]
module = "ib_async.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "legacy/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
